misc: refactor compose file (#17)
##  Summary 

This PR refactor the docker compose file to allow building either using
the local Dockerfile or using the remote git url to the project. It is
useful when you want to only copy past the docker compose file and run
it without the source code. It also pin postgres to version 18 and
update the volume path.

## What changed

- Rename `docker-compose.yaml` to `compose.yml` (shorter name)
- Update README.md and complete the "Quick start" with extra steps
- Clean `compose.yml`, add commented `build` field for remote build file
- Update postgres to 18 and change volume path to the one use in version
18
- Remove migrations... Don't really need it as I am restarting from
scratch.